1525 Extract attachment actions into component
